examples/hashing: Update example application to measure hashing performance of SHA-256 and Keccak
Performance is measured in ticks needed to calculate 1e6 hash operations and in hash operations per tick. The hashes are calculated of random strings of lengths from a given set. For Keccak, there are two different types of benchmark:

1. A benchmark normalized around the number of full state transformations (regardless of state size).
2. A benchmark normalized around the state size.

For each of them a rate is calculated accordingly.

Also included is a function to generate a string with the result of a division without using floating point arithmetic. The string is cut off after a certain precision. If applicable and configured, the last digit is rounded up.